# api/openapi.yaml
openapi: 3.0.3
info:
  title: PingMesh API
  description: |
    REST API for PingMesh — a distributed multi-vantage monitoring system.

    PingMesh deploys lightweight nodes across different networks that run checks
    (HTTP, TCP, DNS, ICMP) from multiple vantage points and use quorum-based
    consensus to confirm failures before alerting.

    ## Architecture

    Each PingMesh node exposes two HTTP servers:

    | Server | Default Port | Purpose |
    |--------|-------------|---------|
    | **CLI API** | `7434` (localhost) | Management API — this spec documents it |
    | **Peer API** | `7433` (all interfaces) | Inter-node communication (heartbeats, config sync, result push) |

    ## Authentication

    The CLI API listens on localhost by default and has **no authentication**.
    If you expose it externally, add a reverse proxy with auth in front of it.

    ## Quick Start

    ```bash
    # List all monitors
    curl http://localhost:7434/api/v1/monitors

    # Create an HTTP monitor
    curl -X POST http://localhost:7434/api/v1/monitors \
      -H 'Content-Type: application/json' \
      -d '{"name": "My Website", "check_type": "http", "target": "example.com"}'

    # Check cluster health
    curl http://localhost:7434/api/v1/health
    ```

    ## Check Types

    | Type | Description | Required Fields |
    |------|-------------|-----------------|
    | `http` | HTTP GET request | `target` (hostname) |
    | `https` | HTTPS GET request | `target` (hostname) |
...

...

...

...

...
servers:
...
paths:
...
  /api/v1/monitors:
    get:
      tags: [Monitors]
      summary: List monitors
      description: Returns all monitors, optionally filtered by group name.
      operationId: listMonitors
      parameters:
        - name: group
          in: query
          description: Filter by group name (exact match)
          schema:
            type: string
          example: production
      responses:
        "200":
          description: Array of monitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Monitor"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Monitors]
      summary: Create a monitor
      description: |
        Creates a new monitoring check. The server generates the `id`, `created_at`,
        `updated_at` fields and applies defaults for any omitted optional fields.

        **Minimum required fields:** `name`, `check_type`, `target`
      operationId: createMonitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorCreate"
            examples:
              http:
                summary: HTTP check
                value:
                  name: "My Website"
                  check_type: "http"
                  target: "example.com"
                  interval_ms: 30000
              tcp:
                summary: TCP port check
                value:
                  name: "Database"
                  check_type: "tcp"
                  target: "db.internal"
                  port: 5432
                  timeout_ms: 3000
              dns:
                summary: DNS resolution check
                value:
                  name: "DNS A record"
                  check_type: "dns"
                  target: "example.com"
                  dns_record_type: "A"
                  dns_expected: "93.184.216.34"
              keyword:
                summary: HTTP keyword check
                value:
                  name: "Login Page"
                  check_type: "http_keyword"
                  target: "app.example.com"
                  expected_keyword: "Sign In"
      responses:
        "201":
          description: Monitor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/monitors/{id}:
    parameters:
      - $ref: "#/components/parameters/MonitorId"

    get:
      tags: [Monitors]
      summary: Get a monitor
      operationId: getMonitor
      responses:
        "200":
          description: Monitor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Monitors]
      summary: Update a monitor
      description: |
        Partially updates a monitor. Only non-zero/non-empty fields in the request
        body are applied. Fields you omit (or set to zero) are left unchanged.

        **Updatable fields:** `name`, `target`, `port`, `interval_ms`, `timeout_ms`, `group_name`
      operationId: updateMonitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorUpdate"
            example:
              name: "Updated Name"
              interval_ms: 15000
      responses:
        "200":
          description: Updated monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Monitors]
      summary: Delete a monitor
      operationId: deleteMonitor
      responses:
        "200":
          description: Monitor deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
              example:
                status: deleted
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Nodes ─────────────────────────────────────────────────────────────
  /api/v1/nodes:
    get:
      tags: [Nodes]
      summary: List cluster nodes
      description: Returns all nodes in the cluster with their current status.
      operationId: listNodes
      responses:
        "200":
          description: Array of nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/nodes/{id}:
    parameters:
      - $ref: "#/components/parameters/NodeId"

    get:
      tags: [Nodes]
      summary: Get a node
      operationId: getNode
      responses:
        "200":
          description: Node details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Nodes]
      summary: Remove a node from the cluster
      operationId: deleteNode
      responses:
        "200":
          description: Node deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
              example:
                status: deleted
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Status & Incidents ────────────────────────────────────────────────
...
  /api/v1/incidents:
    get:
      tags: [Status]
      summary: List incidents
      description: Returns incidents. Use `?active=true` to see only unresolved incidents.
      operationId: listIncidents
      parameters:
        - name: active
          in: query
          description: If `true`, return only active (non-resolved) incidents
          schema:
            type: string
            enum: ["true", "false"]
          example: "true"
      responses:
        "200":
          description: Array of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/history:
    get:
      tags: [Status]
      summary: Check result history
      description: |
        Returns raw check results. All parameters are optional filters.
        Results are returned newest-first, up to `limit`.
      operationId: listHistory
      parameters:
        - name: monitor
          in: query
          description: Filter by monitor ID
          schema:
            type: string
        - name: node
          in: query
          description: Filter by node ID
          schema:
            type: string
        - name: since
          in: query
          description: Only return results after this Unix timestamp (milliseconds)
          schema:
            type: integer
            format: int64
          example: 1700000000000
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 50
            minimum: 1
          example: 100
      responses:
        "200":
          description: Array of check results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckResult"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Alert Channels ──────────────────────────────────────────────────
  /api/v1/alerts/channels:
    get:
      tags: [Alerts]
      summary: List alert channels
      description: Returns all configured alert channels (webhook and email).
      operationId: listAlertChannels
      responses:
        "200":
          description: Array of alert channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertChannel"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Alerts]
      summary: Create an alert channel
      description: |
        Creates a new alert channel. The `config` field must be a JSON string
        containing either a WebhookConfig or EmailConfig depending on `type`.
      operationId: createAlertChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertChannelCreate"
            examples:
              webhook:
                summary: Webhook channel
                value:
                  name: "Slack Alerts"
                  type: "webhook"
                  config: '{"url":"https://hooks.slack.com/services/...","secret":""}'
              email:
                summary: Email channel
                value:
                  name: "Ops Team"
                  type: "email"
                  config: '{"smtp_host":"smtp.gmail.com","smtp_port":587,"username":"alerts@co.com","password":"app-pwd","from":"alerts@co.com","to":"ops@co.com","tls":true}'
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertChannel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/alerts/channels/{id}:
    parameters:
      - $ref: "#/components/parameters/AlertChannelId"

    get:
      tags: [Alerts]
      summary: Get an alert channel
      operationId: getAlertChannel
      responses:
        "200":
          description: Channel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertChannel"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Alerts]
      summary: Update an alert channel
      description: Update channel name, config, or enabled status.
      operationId: updateAlertChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertChannelUpdate"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertChannel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Alerts]
      summary: Delete an alert channel
      description: Deletes the channel and all associated delivery history (CASCADE).
      operationId: deleteAlertChannel
      responses:
        "200":
          description: Channel deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
              example:
                status: deleted
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/alerts/channels/{id}/test:
    parameters:
      - $ref: "#/components/parameters/AlertChannelId"

    post:
      tags: [Alerts]
      summary: Send a test alert
      description: |
        Sends a synthetic test alert through the specified channel.
        The result is recorded in alert history with event_type "test".
      operationId: testAlertChannel
      responses:
        "200":
          description: Test sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
              example:
                status: sent
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: Alert dispatcher not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/alerts/history:
    get:
      tags: [Alerts]
      summary: Alert delivery history
      description: |
        Returns alert delivery records (success and failure) ordered newest-first.
        Optionally filter by channel ID and limit results.
      operationId: listAlertHistory
      parameters:
        - name: channel
          in: query
          description: Filter by channel ID
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum records to return
          schema:
            type: integer
            default: 50
            minimum: 1
          example: 100
      responses:
        "200":
          description: Array of alert records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertRecord"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Health & Diagnostics ──────────────────────────────────────────────
...
  /api/v1/test-peer:
    get:
      tags: [Health]
      summary: Test peer connectivity
      description: |
        TCP-dials each peer node (or a specific one) and returns reachability
        status with latency. Useful for diagnosing network issues.
      operationId: testPeer
      parameters:
        - name: node
          in: query
          description: Test only this node ID (omit to test all peers)
          schema:
            type: string
      responses:
        "200":
          description: Array of peer statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeerStatus"
  /api/v1/logs:
    get:
      tags: [Logs]
      summary: Recent log entries
      description: |
        Returns the most recent log entries from the in-memory ring buffer
        (capacity: 1000 entries). Entries are returned oldest-first.
      operationId: getLogs
      parameters:
        - name: lines
          in: query
          description: Number of log lines to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          example: 50
      responses:
        "200":
          description: Array of log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
              example:
                - time: "2026-02-17T22:07:44.844Z"
                  message: "[agent] starting node abc-123 (prod-us-east) role=coordinator"
                - time: "2026-02-17T22:07:44.912Z"
                  message: "[check] Google HTTP → up (66.4ms)"
        "503":
          description: Log buffer not available (agent started without log buffer)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

# ═══════════════════════════════════════════════════════════════════════════
# Components
# ═══════════════════════════════════════════════════════════════════════════
components:
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "database error"
  schemas:
...
    Monitor:
      type: object
      description: A monitoring check configuration
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier (server-generated)
          example: "a0373a66-ba17-4e93-985d-a1865890036a"
        name:
          type: string
          description: Human-readable name
          example: "Google HTTP"
        group_name:
          type: string
          description: Optional group for organizing monitors
          example: "production"
        check_type:
          type: string
          description: Type of check to perform
          enum: [icmp, tcp, http, https, dns, http_keyword]
          example: "http"
        target:
          type: string
          description: |
            Target to check. For HTTP/HTTPS: hostname without scheme (e.g. `example.com`).
            For TCP: hostname or IP. For ICMP: IP or hostname. For DNS: domain to resolve.
          example: "www.google.com"
        port:
          type: integer
          description: Port number (required for TCP checks, optional for HTTP/HTTPS)
          example: 443
        interval_ms:
          type: integer
          format: int64
          description: Check interval in milliseconds
          default: 60000
          example: 30000
        timeout_ms:
          type: integer
          format: int64
          description: Timeout per check attempt in milliseconds
          default: 5000
          example: 5000
        retries:
          type: integer
          description: Number of attempts before marking as failed
          default: 1
          example: 1
        expected_status:
          type: integer
          description: Expected HTTP status code (HTTP/HTTPS checks only)
          example: 200
        expected_keyword:
          type: string
          description: Keyword to search for in response body (http_keyword check only)
          example: "Sign In"
        dns_record_type:
          type: string
          description: DNS record type to query (dns check only)
          enum: [A, AAAA, CNAME, MX, TXT, NS, SOA]
          example: "A"
        dns_expected:
          type: string
          description: Expected DNS response (dns check only)
          example: "93.184.216.34"
        failure_threshold:
          type: integer
          description: Consecutive failures before a node reports the target as down
          default: 3
          example: 3
        recovery_threshold:
          type: integer
          description: Consecutive successes before a node reports recovery
          default: 2
          example: 2
        quorum_type:
          type: string
          description: |
            How many nodes must agree for consensus.
            - `majority`: more than half of online nodes
            - `n_of_m`: at least `quorum_n` nodes
          enum: [majority, n_of_m]
          default: "majority"
          example: "majority"
        quorum_n:
          type: integer
          description: Required agreeing nodes when `quorum_type` is `n_of_m`
          example: 2
        cooldown_ms:
          type: integer
          format: int64
          description: Minimum time between repeated alerts for the same monitor (ms)
          default: 300000
          example: 300000
        enabled:
          type: boolean
          description: Whether this monitor is active
          default: true
          example: true
        created_at:
          type: integer
          format: int64
          description: Creation timestamp (Unix milliseconds, server-generated)
          readOnly: true
          example: 1771364388838
        updated_at:
          type: integer
          format: int64
          description: Last update timestamp (Unix milliseconds, server-generated)
          readOnly: true
          example: 1771364388838
...